@@ -17,23 +17,18 @@ namespace Hirieasu_Andrei_Lab2.Data
                         return; // BD a fost creata anterior
                     }
                     context.Books.AddRange(
-                    new Book{Title = "Baltagul",Author = "Mihail Sadoveanu",Price=Decimal.Parse("22")},
+                    new Book{Title = "Baltagul",Price=Decimal.Parse("22")},
                    
-                    new Book {Title = "Enigma Otiliei",Author = "George Calinescu",Price=Decimal.Parse("18")},
+                    new Book {Title = "Enigma Otiliei",Price=Decimal.Parse("18")},
                    
-                    new Book{Title = "Maytrei",Author = "Mircea Eliade",Price=Decimal.Parse("27")}
+                    new Book{Title = "Maytrei",Price=Decimal.Parse("27")}
                    
                     );
 
 
                     context.Customers.AddRange(
-                    new Customer 
-                    { 
-                        Name = "Popescu Marcela",Adress = "Str. Plopilor, nr. 24", BirthDate = DateTime.Parse("1979-09-01")
-                    },
-                    new Customer
-                    {
-                        Name = "Mihailescu Cornel", Adress = "Str. Bucuresti, nr. 45 ,ap. 2",BirthDate=DateTime.Parse("1969 - 07 - 08")}
+                    new Customer {  Name = "Popescu Marcela",Adress = "Str. Plopilor, nr. 24", BirthDate = DateTime.Parse("1979-09-01")},
+                    new Customer{ Name = "Mihailescu Cornel", Adress = "Str. Bucuresti, nr. 45 ,ap. 2",BirthDate=DateTime.Parse("1969 - 07 - 08")}
                    
                     );
 
