@@ -18,9 +18,11 @@
                 <span asp-validation-for="Title" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="Author" class="control-label"></label>
-                <input asp-for="Author" class="form-control" />
-                <span asp-validation-for="Author" class="text-danger"></span>
+                
+                <label asp-for="ID" class="control-label"></label>
+                <select asp-for="ID" class="form-control" asp-items="ViewBag.ID"></select>
+               @*<input asp-for="Author" class="form-control" />
+                <span asp-validation-for="Author" class="text-danger"></span>*@
             </div>
             <div class="form-group">
                 <label asp-for="Price" class="control-label"></label>
